Temp commit

+++ app/src/main/java/com/example/revolut/data/CurrencyResponse.kt
@@ -5,5 +5,5 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class CurrencyResponse(
     val baseCurrency: String,
-    val rates: Map<String, Double>
+    val rates: Map<String, Float>
 )
